Let's support 64 lights per scene instead of 32.

+++ src/client/renderer/r_main.h
@@ -102,8 +102,8 @@ typedef struct r_locals_s {
 
 	int16_t light_frame; // dynamic lighting frame
 
-	uint32_t active_light_mask; // a bit mask into r_view.lights
-	uint32_t active_light_count; // a count of active lights
+	uint64_t active_light_mask; // a bit mask into r_view.lights
+	uint16_t active_light_count; // a count of active lights
 
 	c_bsp_plane_t frustum[4]; // for box culling
 }r_locals_t;
